增加viewModle.go定义view包中结构体，用来对接前后端json
